# Make babel-types type checking functions 36% faster