When profiling babel, I found that the type checking functions in `babel-types` were the slowest part of the code by far. They are also some of the hottest functions in the code base, called a ton.

Previously, they would need to call two layers of functions, do several object lookups to get alias keys, and then loop through all of those to check them. This PR inlines all of that into the pre-generated functions so it's basically only a single if statement.

I benchmarked by running `babel-preset-minify` over the lodash source code (527K). This change made that roughly 25% faster, bringing the minification time from ~4570ms to ~3515ms. (Edit: See below, now 36% faster due to an additional improvement)

Here's the code I ran using the current `@babel/core` version on npm vs the code in this change.

```javascript
const babel = require('@babel/core');
const minify = require('babel-preset-minify');

babel.transformFileSync(__dirname + '/node_modules/lodash/lodash.js', {
  presets: [minify]
});
```

And here's the full output of the benchmark:

```
original x 0.25 ops/sec ±3.30% (5 runs sampled)
new x 0.34 ops/sec ±5.12% (5 runs sampled)
original is 25% slower
new is the fastest
```