@devongovett 2018-04-07
Added an additional commit to pre-fetch the type checking functions from babel-types when generating the corresponding NodePath methods. This adds an additional 11% speed improvement to the same benchmark for a total of 36%. Minification time for lodash is now 3202ms, vs 4570ms before this PR.

It was slow because the object lookup actually resolves to a function call to require babel-types on demand. We can avoid this and pre-cache the results of the object lookup as well.

New benchmark output:

```
original x 0.26 ops/sec ±5.58% (5 runs sampled)
new x 0.41 ops/sec ±2.89% (6 runs sampled)
original is 36% slower
new is the fastest
```